+++ src/styles/index.ts
@@ -1,3 +1,14 @@
-const frontendStylesheet = "./styles.css";
+const frontendStyleSubpaths = Object.freeze([
+  "./styles/tokens",
+  "./styles/utils",
+  "./actions/styles",
+  "./flash/styles",
+  "./progress/styles",
+  "./layer/styles",
+  "./tooltip/styles",
+  "./popover/styles",
+  "./modal/styles",
+  "./inputs/styles",
+]);
 
-export { frontendStylesheet };
+export { frontendStyleSubpaths };